+++ js/components/setting_button.js
