+++ js/scenes/open_logo.js
